requirements.txt
numpy
pandas
tensorflow
scikit-learn
flask
python-dotenv
